[packet-formats] Allow updating raw ports/addrs
As part of this, update the header checksum (if present). We don't know
if the checksum is valid, but we at least maintain that any updates
don't make it *more* invalid :-)

This will be used as part of NATing ICMP errors, where we want the inner
packet to have the "right" checksum even though we might not have the
whole packet in order to validate it.

Bug: 341340810
Change-Id: Ib1afd562cf55a5102b3ac4b21ec26e0f32bc3028
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255405
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter DeLong <pdelong@google.com>