[zx] Expose signals from zx_object_wait.
Currently, wait and handle_wait in the zx crate only return Signals when
the syscall result is ZX_OK. However, the docs make it clear that the
result is defined when the status is ZX_OK, ZX_ERR_TIMED_OUT, or
ZX_ERR_CANCELED.

ZX_ERR_TIMED_OUT is useful in practice: it's common to wait with an
INFINITE_PAST deadline in order to see the signals currently asserted on
an object.

ZX_ERR_CANCELED is less likely to be useful, but since it's defined
according to the docs, we should probably expose it at this layer.

Bug: 410869980

Change-Id: I469fd4ea4a64dbfaa76f6feae6dbd1ac89a08647
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253204
Fuchsia-Auto-Submit: Hunter Freyer <hjfreyer@google.com>
Owners-Override: James Robinson <jamesr@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>