[ffx] Make new target list JSON schema compatible with old
The introduction of non-IP address types changed this schema, and we
keep tripping over migration pains. This changes the schema again to
something that looks more like the old one and should let tools ignore
it if they haven't been updated.

The places where we attempted to migrate all had to keep backwards
compatibility with the old schema so this should be able to land safely.

In the future we should find a way to make changes to the schema goldens
block changes harder. As it stands they're just golden files that in a
lot of engineers' build configurations will just automatically update if
you change the schema.

Change-Id: I0dde02e6f2ea207c11c1ddd3a2d4a47ca435d968
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255852
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>