[kernel][heap] Trace how long growing the heap takes
This CL adds a trace duration that captures how long it took to grow
the heap during an allocation and how many retries it took to do so.

It also slightly refactors the calculation of `growby` to make it
easier to understand.

Bug: 399427107
Change-Id: I4a399cccfbbe12754b0eb737b981ebdce863a105
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253439
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>