[kernel] Implement KTrace::ReportMetadata for per-CPU mode
Implements the ReportMetadata method for KTrace when running in per-CPU
mode. This is a combination of the ReportStaticNames and
ReportThreadProcessNames methods in internal::KTraceState.

This CL also updates the tests to verify that the metadata is emitted
exactly once when tracing is started. This required overriding the
ReportMetadata method in a derived class because the thread and process
names are written into the global singleton's trace buffer, which we do
not want to manipulate in tests.

Bug: 404539312
Change-Id: I1749fce2ec585ea0b412373fc9092ca3e8acca8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1250804
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>