[netstack3] Concretize ICMP message functions
This is required for ICMP error NATing because otherwise we end up
with trait bounds that are impossible to meet without conflicting
implementations. For example, needing to implement something for
`Nested<Concrete, IcmpMessage<I, Generic>>` as well as
`Nested<Generic, IcmpMessage<I, Concrete>>`.

This will theoretically have a minor performance impact, but none of
these codepaths are expected to be anywhere close to hot, since these
should only be called for errors or NDP.

Bug: 328057704
Change-Id: Id4accf40629cd4654d49ae6c14377b17f90af81c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255145
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>