[kernel] Move FXT metadata record writes into ReportMetadata
This patch moves writing the magic and initialization records from the
Rewind method into the `ReportMetadata` method for the per-CPU
implementation of KTrace.

Mechanically, this means:
1. Using the fxt library methods to emit the magic and initialization
   records instead of hardcoding their values inline, and
2. Using an IPI to emit these records on the boot CPU so that they're
   always read first during a read.
3. Moving the implementation of ReportMetadata from the header to the
   CPP file, which required moving the Reserve function ahead of it to
   appease the template instantiation requirements.

Note that this does slightly change the behavior of ktrace; the magic
and initialization records are emitted every time a new ktrace session
is started, instead of just when the buffer is initially allocated or
rewind'ed.

I've tested that this works by taking a trace locally; future patches
will rework `ReportMetadata` to make it more testable so we can exercise
it in unit tests.

Bug: 404539312
Change-Id: Ia7257c8427caa7d8b1a6e388d20cb5ca8fad9f4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257090
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>