[fidl][rust] Moveable decoded wire values
This change introduces the `Decoded` type as a variant of `Owned` which
does own the backing memory, and provides methods on `DecoderExt` to
decode into a `Decoded`. To that end, this change also introduces the
`commit` method on `Decoder` which allows resources to be dropped even
when decoding fails part of the way through.

The API usage examples and benchmark have been updated to reflect the
new decoding API.

Change-Id: Iaae913ee7353a1371ae3da48dae3b62e07284aea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253165
Reviewed-by: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: David Koloski <dkoloski@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>