[reachability] Downgrade fetch and ping to INFO
Change WARN logs to INFO logs to clarify that these logs may not be an
issue depending on environment setup. For example, ping failures over
IPv6 are expected if a device is connected to a IPv4-only network.

Log lines will be reduced or removed fully when migrating to a new
inspect data format.

Bug: 42085065
Change-Id: I042743252e0b5a4ae7f44c6feaf6a9403f34c39d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257846
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>