[block_server][c] Track request state
The server keeps track of active request state so the C++ side doesn't
need to be trusted to do so.  The C++ side now use request IDs which
are indices into the internal request state map, and respond to requests
by ID.

This was necessary because C clients which asynchronously respond to
messages don't work with the current API, where BlockServer::OnRequests
passes in a const Session& (expecting a synchronous call to
Session::SendReply before OnRequests returns).  The client can't just
store the Session as a pointer because this is a temporary wrapper
object.  One option was to create a ref type (e.g. SessionRef) and pass
that in, but there is a greater change of use-after-free type issues
with that, versus the approach taken here.

This has several other advantages:
- Internally we don't need to use high-bits in request IDs to decide
  whether to discard the VMO on request completion; the internal request
  state keeps a strong reference to the VMO if needed.
- We were able to get rid of the hack to create a Session that didn't
  run its destructor in block_server.cc.
- The C API becomes a bit simpler.

Bug: 410926221
Change-Id: If921f4124c9ba8295ff8d6c1a67ed0a92093d345
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255385
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>