[netdevice] Fix crash when switching sessions
Observed when adding controlled shutdown to netstack3. There's a
specific series of events that cause the attempt to take over primary
session when a session newly attaches a port that attempts to insert a
nullptr in a doubly linked list.

Fix the code with a simple check and add a regression test.

Change-Id: I1c94d92ff30e23d5757a84e4d4f97321ed11bb6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1259365
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>