[tracing][fidl] Add defer_transfer TraceConfig Option
Currently, in streaming mode, trace manager attempts to write trace data
as it receives it to a socket back to whomever asked it for data. This
is fine when the consumer of the trace in on device, it's just a socket
write. However, in the case of ffx, the other end is on the host and
this triggers a fair amount of overhead sending data over netstack/usb.

This adds a flag which we can propate through trace manager to hold off
on transferring the file (buffer it to disk) until the trace if over.

Bug: 413106095
Change-Id: I7068e09c4f478403f84f36676a61f65e85d958e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260406
Reviewed-by: Krithika Balan <krithikabalan@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
API-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>