[netstack3] Fix reassembly of IP fragments
Fix a bug in the reassembly of out-of-order IP fragments. This original
code was incorrectly computing the gaps between fragments, only when 1)
the current fragment fills a gap *before* an existing fragment (i.e. an
out-of-order fragment) and 2) when the current fragment's size is
greater than `FRAGMENT_BLOCK_SIZE`.

The fix itself is trivial; the majority of this diff in this CL is to
update the tests to support testing reassembly on fragments of different
sizes. The original test coverage only exercised fragments with
`FRAGMENT_BLOCK_SIZE`, which kept this bug hidden.

Bug: 413346888
Change-Id: I1bc3cbc509f6c74f96c5c5156191785570e65b40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260604
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>