[fshost] mock boot instead of falling back to pkg
The zxcrypt configuration exists in /boot/config at runtime. For the
integration tests, we included a config/zxcrypt in the test fshost
package, so the path was /pkg/config/zxcrypt. To work around this,
get_policy had a fallback mechanism.

There is no reason to do it like this anymore. We can mock /boot in the
tests to provide whatever the test would like. Now we provide a read
only file for the zxcrypt policy, and remove the fallback. The only
thing that used get_policy to read the policy file was fshost.

This also removes the x permission from boot. I'm pretty sure we don't
execute anything from boot anymore now that we are a bootfs package -
all the binaries we need are either included in our package or resolved
as packages via the boot resolver. This is verified by CQ and by the
fact that we were never giving any form of /boot to fshost in the
integration tests so it's apparently not that important - I think it's
only used for the zxcrypt policy right now.

Change-Id: I1fa4023251124f4b69511976ea030a9db5aa2bfb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260468
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>