[starnix] Promote the Socket's FsNode to a strong reference
Maintaining a weak reference to the Socket's FsNode is unnecessary
since the only reference from FsNode to Socket is from the file-
system node to which a Unix-domain Socket is bound, to the Socket,
and in that case the FsNode differs from the one associated with
the Socket itself.

Holding a strong reference avoids raciness between referencing the
FsNode of a target Socket, and it being torn-down by its owner.

Change-Id: I2d1e8b736cc88d86bcd30022298c0c43e5208ab0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261264
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>