[idk][build] api_level is always a string in IDK metadata
https://fxrev.dev/1082028 added support for strings while keeping
numbered API levels as integers. Based on experience with this in an
SDK, it will be simpler if the level is always a string. The value is
mostly used to build paths, so this shouldn't be a problem. Anything
that needs the integer value of a level should get them from
version_history.json.

Bug: 326277078
Change-Id: Ie728aba6787ec013a8edc9f69f159dfa5f8254fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087174
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>