[kernel][arm64] Restructure harvest lock dropping
Restructures the harvesting code to only drop the lock if the lock is
actually contended, otherwise reset the entry counter. The top level
function is also rearranged to avoid the use of CallUnlocked, which can
lead to errors with the ConsistencyManager, and avoid using a busy loop
when waiting for mark accessed threads and blocking on the mutex
instead.

Resetting the entry_limit, instead of bouncing out and back into the
harvest loop, reduces the average runtime of harvesting on devices by
up to 40%.

Change-Id: Id65a50e8745b58ab24830d45d5550b88b54707d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1248544
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>