[fxfs] Fixed duplicate tombstone issue
Prior to this CL, if you were to create an unnamed temporary file, and
then close it without linking it anywhere, some time later when the
timed flushing occurs, the file would get tombstoned for a second time.

The root cause if this issue is that prior to this CL, we would allow an
open count to be taken on a file that had already been purged if it was
an unnamed temporary file.

The fix is to disallow this, but in order to do so, we need to change
how the file is created.

This also fixes an issue where we errnoneously took the truncate guard
when queueing a tombstone, instead of when we actually executed the
tombstone.

Bug: 412830759
Change-Id: Ic4c4622172f934679bd8b83cb8546dcd914af67f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1259844
Reviewed-by: Rui Qi Sim <rqsim@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>