[fx] Remove FUCHSIA_BUILD_DIR from fx.config file.
This fixes an issue that happens when renaming a build directory,
`fx --dir out/newname build` would continue to build in òut/oldname
because it was stored in the fx.config file.

There doesn´t seem any good reason to write the build directory in
that file, since it is only read by `fx` after the build directory
was already set (e.g. with the --dir option, or reading the
.fx-build-dir value, which is always correct).

Fixed: 413713238
Change-Id: I46c0f947ba52f33637908ebbff38a02a311cb808
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261405
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>