[starnix][ext4] Move pager out of starnix kernel
When suspending a Starnix container all the threads in the container are
paused. If the pager threads are suspended before the pager client's the
suspend operation may deadlock. This is because the client threads can
end up blocked in Zircon (on the pager response), which prevents them
from being suspended.

This change moves the pager threads into the starnix_runner component,
which is the same component that is responsible for suspending the
container threads. This guarantees that the pager requests can always be
handled.

Fixed: 410933064

Change-Id: I2bbb6b3aeedffbdb4a368715dcf36357b4f5f490
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260704
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>