[memory] Factor out testing code.
It is not possible to depend on the code gated by `#[cfg(test)]`
from another crate. Having this code in a separate crate would
require to define the trait in one crate, the testing code in
another, and have memory/attribution/processing depend on both.

This solution is simpler, and the struct will be stripped out
by the linker.

Change-Id: Ib92209ea7935b03faa6ba2bc77684f866e562bb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262368
Fuchsia-Auto-Submit: Pierre Labatut <plabatut@google.com>
Reviewed-by: Fabio D'Urso <fdurso@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>