[fxfs] Use Arc::new_cyclic with the pager
PagerPacketReceiver required 2 phase construction because the receiver
needs a Weak of the file and file needs to own the receiver.

Arc::new_cyclic was stabilized a while ago which gives out a Weak to the
file before the file is constructed. This allows the receiver to be
constructed without needing an uninitialized state and to be initialized
later.

Change-Id: If59c9ddb998706c26a8fcd8d373e2374d1d1c8cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261164
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>