[fidl][ffx] RemoteControl is server=platform
Mark the RemoteControl protocol as server=platform to allow easier
evolution. This allows us to add the host_tool category of FIDL to the
abi checker. The definition of "platform" in the host tool context is
weird and ambiguous but the remote control service is definitely in
it...

Bug: 413661651
Change-Id: I2289876518dddb9a026b65c4b4350ba5b5469921
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261484
Reviewed-by: Casey Dahlin <sadmac@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>