[netstack3] Bake ICMP mutability into trait
Now that there are conntrack entries associated with ICMP errors, NAT
will attempt to rewrite the ID of the outer ICMP message if it's an
error and there's a valid inner payload.  Previously, the code assumed
this was impossible, so update_icmp_id will panic.

To resolve the issue, this change adds an IS_REWRITABLE constant to
the IcmpMessage trait that can be checked before the code attempts to
rewrite the ID.

Bug: 328057704
Change-Id: I0ba7b2947ab3d5f0784d83c7bd695994e1705011
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262884
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter DeLong <pdelong@google.com>