[ffx] Use scope ID strings for ssh.
Scope ID here refers to a link-local IPv6 address scope ID.

Instead of running ssh commands with scope ID's as integers, use
verified scope ID strings.

This adds a `ScopedSocketAddr` struct to be used in ssh and related
commands. This structure is intended to format a stable scope ID in the
form of a string. It also, when constructed, will verify the
scope ID of the address, returning an error if it cannot be found.

The approach taken here is to prevent having to rewrite the majority
of ffx daemon related code. It is not necessarily the most complete.
Inside the daemon is several layers (in several directions) of
address structs: TargetAddrInfo, TargetIp, TargetIpPort, TargetInfo,
TargetAddr, etc. Some of these are directly useful as they abstract
things like vsock connections, but most have dubious usefulness,
especially with the advent of moving more functionality away from the
daemon. Ideally this `ScopedSocketAddr` would be pushed closer and
closer to the first invocation of the user, therefore preventing invalid
scope ID's from making it to the `ssh` command invocation.

Bug: 411230336

Change-Id: I5511fe6f8f0fdd338730554875cd417bd7d1b04c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260729
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>