[tracing] Rename instant_flow_* to instaflow_*
Matches the newly-added C++ macro names.  Since the C++ names are part
of the public SDK, we decided that it was better to pick the name that
we preferred (since it would be difficult to change later), and make the
Rust macros match.

Fixed: 413095076
Change-Id: I765ebcff8e3c16bf7ceabbefe37fd18f1ef04281
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260824
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>