[fxfs] Fix profile tests. Shouldn't be marking as purged.
These were being marked as purged to receive an open count and verify
that it was zero. Not only was this not a sure way to check, but marking
an object to be purged without adding it to the graveyard isn't safe if
the purge actually happens, it can result in a partially deleted object.

Change-Id: Ib8128dd6427e2d311045b8781664109002072d12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1263964
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>