[reachability] Differentiate errors for ping/fetch
It is normal for errors from the ping and fetch protocol to either be
network unreachable or host unreachable. This does not always signify
a networking issue, it could be a network configuration where IPv6 is
disabled, for example. In this case, log the error messages at INFO
instead of WARN. WARN errors will be classified as errors that may have
underlying networking problems.

Bug: 42085065
Change-Id: I053523b8f6289e02a0044319be3b3937127c2404
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262491
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>