[io] Workaround for remove-and-rename.
fuchsia.io.* related Deprecated methods aren't quite correct.

For example this:

    @available(added=25, removed=26)
    @selector("Open")
    strict DeprecatedOpen(resource struct {

This will keep DeprecatedOpen around if and only if API level 25 is
supported, but that's not what we want. We want it to be supported (for
compatibility purposes) as long as any API level that supported Open is
supported (that is, any between 11 and 25).

We don't have a nice way of doing that right now. This CL changes them
to work in a variety of _different_ bad ways, but ways that allow us to
retire unused API levels without things breaking.

Bug: 324080864
Bug: 384994764

Change-Id: I36b320687129806c121fff9bf536e0a561ebe729
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255651
API-Review: Hunter Freyer <hjfreyer@google.com>
Fuchsia-Auto-Submit: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>