[netstack3/filter] Add ICMP payload modification
This is the bulk of the mechanism required to support NAT for ICMP
errors, where we want to be able to rewrite the inner IP and transport
payloads in addition to the outer IP header.

The basic idea of the implementation is that we return something that
implements IpPacket on the inner payload to allow modification just like
normal IP and transport headers.

Also catch a few more places where FilterIpExt needed to be plumbed.

Bug: 328057704
Change-Id: I2e5b4b5b0df91c4835d5a8fbace40758e83d09e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255148
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>