[block_server] Split large requests
Split large requests in the "core" block server before they make it to
the concrete block server implementation.

Large requests are placed into an ad-hoc "one-shot" group, to ensure
that only one reply goes out on the FIFO.  (For requests that are
already in a group, we simply accumulate the additional requests into
the existing group).

I opted to split requests at the same time that we check the OffsetMap
for remapping requests, because we will also need to split requests for
discontiguous requests.

Also adds the ability to set maximum_transfer_size to the ramdisk, and
fixes a bug (TRIM sends two responses).  This is used for a C++
integration test.

Fixed: 372555590
Change-Id: I3f8730c38361e867b63a42d3f05ac06a8c8fe76f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261608
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
API-Review: Chris Suter <csuter@google.com>