[kernel][arm64] Clean kernel load image before secondary CPUs start
This is a step towards removing set-way data cache operations, but also
to finally put to bed the previous problem of how to ensure that
pre-cache secondary CPUs have a view of the relevant memory that's
coherent with the boot CPU. As reasoned elsewhere, it should suffice to
ensure that (1) the code the secondary CPUs would touch before enabling
caching and (2) the variables it loads are cleaned to the point of
coherency. While we could be surgical about that, it suffices to simply
clean the whole kernel load image, which surely would include (1) and
(2).

Previously this was done with an even larger hammer of cleaning the
whole boot CPU data cache. We don't need to go that big - and our desire
to not use set-way data cache operations suggests that we need an'
alternative. Also, for historical reasons, this call was made in
physboot; now it's done just before secondary CPUs are brought online
(i.e., at the top of topology_cpu_init()).

Also removed in this change are the artifacts of the previous attempt to
be more surgical, as alluded to above. The pitfall with that attempt was
that we didn't end up actually cleaning all of the secondary code! While
we properly bounded the instructions and cleaned that range, we didn't
account for branches taken within.

Bug: 408020980
Change-Id: Iff157360a3501c2b4c1bff2b5f71ddbed1c320a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261588
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>