[logs] Remove explicit shard reference in RealmBuilder shard
The RealmBuilder shard is shipped in the SDK.
SDK consumers may want custom routing for LogSink
and InspectSink. Forcing them to include the shard
removes a lot of flexibility, and makes migrations
of LogSink routing more difficult, not simpler.
All out-of-tree users have migrated to explicit
routes or explicit dependencies on the LogSink shard,
so it's safe to update this.

Fixed: 404543928
Change-Id: I8b967811c47c74c4429d71e493d81db21fbeda16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1258026
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Christopher Johnson <crjohns@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>