[time][tests] Convert some tests to run in fake time
Quite a few of timekeeper tests have been written to run in real time.
Unfortunately, this means that the tests are sensitive to adding more
coroutine work, or changing the task scheduling. Both lead to flakes.

I noticed a few tests that were failing this way locally when I fixed
the way initial sampling delay is treated, and I converted those tests
to run in fake time.  The main difference is that we no longer run until
the event loop is stalled, which is a fickle criterion. Instead, we run
the event loop for a predefined length of (fake) time.

This fix is not directly related to b/404936065, but should make the
upcoming changes test more predictably.

Bug: 404936065
Change-Id: I972d70d43b6b7cf4c60d8dcaae2cdca4d5ffd7b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1263826
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>