[audio][test] Make VirtualAudioDriver compliant
VirtualAudioComposite instances are now detectable by the
AudioDeviceRegistry service, but ADR reveals that the
default configuration for Composite does not comply with
requirements for the `signalprocessing` protocol.

By default VirtualAudioComposite exposes a RingBuffer routed
to a DAI. The DAI does not set `plug_detect_capabilities` as
required, and in response to `WatchElementState` neither of
the two elements specify the required field `started`.

This CL fixes those omissions, making the default config
compliant. This is a prerequisite for VAD to be used as a
test fixture, to hermetically test ADR and its clients.

We also update virtual_audio_util so that 'Composite' is the
default assumed driver type. While we are here, clean up any
clang-tidy issues in this source file.

Bug: 392662238

Change-Id: I94ecc6e9152410719d57ccc3b110d17c9b9ffe05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261446
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>