[botanist] Do not powercyle NUCs in fastboot during retry.
Powercycling NUCs in fastboot will cause the iPXE chain loader to run
and take the devices out of fastboot momentarily. Also, the check for
presence of `USB RESET` in serial logs after powercylcing will fail
because this is not part of the logs when iPXE is booting the NUCs into
fastboot.

We can rely on DMS to get the NUCs into fastboot. Botanist should simply
retry to flash or boot the NUC.

Bug: 346868967
Change-Id: I909dcfacffcc8574fdac6480a4a3b9fab902b2f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1265284
Commit-Queue: Betram Lalusha <betramlalusha@google.com>
Reviewed-by: Vinicius Felizardo <felizardo@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>