[component_manager] use the route target when logging routing errors
We were using the component at which a route failed to route errors
about the logging, whereas we want to use the component that started the
routing operation.

This change accomplishes that by relying on the target in the route's
metadata when an error occurs.

Bug: 411475421
Change-Id: Iec4dbb2de9feae395ab5df784c1d4f48975d5aee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262406
Reviewed-by: Olivia Nordquist <nolivia@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>