[idk] Make host field in manifest a proper Clang triple
Previously, the "host" field in the final meta/manifest.json was only a
triple when building on macOS. Insert "unknown-" in other cases so that
it is always a proper Clang target triple. See
https://clang.llvm.org/docs/CrossCompilation.html#target-triple.

This will ensure there is no difference when building the IDK with
Bazel, where `to_clang_target_tuple()` is used.

The string for the most common host changes from "x86_64-linux-gnu" to
"x86_64-unknown-linux-gnu".

Bug: 413071161
Change-Id: Idd4c81dcfd77237ec87b3ad7a935a88423ed5032
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1265006
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>