[time] Fix user time adjust test
I realized that this test is sensitive to the discrepancy between
monotonic timeline and boot timeline. This leads to different test
results on devices that spent different amount of time suspended.

This change removes the issue for this particular test. But there
are other tests with the same problem, which also need fixing. I
am leaving them for a separate change.

Bug: 404936065
Change-Id: Ic99507ad52df120a336e23555c1584a159d14eb8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1265231
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>