[fxfs] Support allow_existing in BlobCreator.
The allow_existing flag will not fail the request if the object already
exists. This also changes edge case behaviour where two writers that
start writing together will both see success when they finish rather
than the second receiving a failure on the final write.

Swaps out the backing for the serving vmo when a blob is replaced while
open. This is done by using an Arc on the registration so that it can be
attached to both the old and the new blob before moving the ownership to
the new blob by pointing the registration at it and adjusting the open
counts to reflect that it has taken ownership of the vmo.

Bug: 409839447

Change-Id: I5776ee7cb5e1afd985096555ab8b39d99e437c8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1263965
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>