[sestarnix] Remove conflicting gtest helper
The starnix test helper defines `SyscallSucceeds`.
The sestarnix test helper defines its own `SyscallSucceeds`,
with a slightly different API.

The 2 definitions conflict, preventing the use of the starnix
test helper in the sestarnix tests.

This CL removes the sestarnix version of `SyscallSucceeds`.
As a cleanup it also removes the sestarnix version of
`SyscallFailsWithErrno`.

Bug: None
Change-Id: I78ca747efafefc02cf381361663c7beb578a9b2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1264464
Commit-Queue: Jean-François Geyelin <jif@google.com>
Reviewed-by: Ambre Williams <ambre@google.com>