[network/testing] update expectations to flaky
The network conformance case 7.4 in the ip/ip_router suites verifies
that the fragmentation cache sets timeouts based on the TTL of the
fragment. This is currently not supported in NS3.

This ultimately leads to a race between the test sending the final
fragment, and the netstack timing out the existing fragments in the
cache. Once we implement this feature the race should be gone.

Bug: 414413500
Change-Id: Iddca6b049a2722a77c54faed0590c278fb2cf2f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266664
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>