[fuchsia-component] Add helpers for custom service name
This change introduces helpers for components that connect to services
that have been given a custom name, and so cannot use the service marker
to auto-fill the desired capability name.

These helpers are to be used, for example, with the proprietary TEE
service implementation in which a service with a proprietary name is
routed through the component topology under a platform-defined name.

Bug: b/410521972
Change-Id: I4d8374c796f8c44e915b760e9151d50fed7035bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266744
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Mark Dittmer <markdittmer@google.com>
Fuchsia-Auto-Submit: Mark Dittmer <markdittmer@google.com>