[netstack3] Limit default route interface watching to main table
Only update has_default_route in the interfaces watcher if we are adding
or removing a route to the main table.

This avoids a bug where we indicate that there is no longer a default
route at all when in fact we've merely removed the last default route in
the table (while other default routes are installed in other tables).

Bug: 412695202
Multiply: interfaces-fidl
Change-Id: I4a4b21ffa4ec9ceb27a42bb72599a4238b851092
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266945
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>