[fidl][rust][fdomain] Build rule cleanup
This is a clean-up of how we build the flavours of Rust FIDL
bindings in our glorious FDomain world. It moves all of the logic and
policy (other than a little backwards compat stuff) from
//build/fidl/fidl.gni to //build/rust/fidl_rust.gni where it belongs.

As a side effect this makes the "common" create for FIDL libraries
depend on the "common" crate for their dependencies rather than on the
main end-user FIDL bindings crate. I noticed that these were showing up
in build trace critical paths. In a local test it's cutting a couple of
seconds off the hot path for a clean ffx build.

Change-Id: I7cb82ef59ab11417ceeb5aec21eb2a4724dc05ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1260905
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>